% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dusoswa_OMIP_54_markers}
\alias{Dusoswa_OMIP_54_markers}
\title{Dusoswa_OMIP_54_markers}
\format{
A data frame with 8 rows and 2 variables:

\describe{
\item{Cluster}{Cluster the cell was grouped into (1-9)}
\item{Markers}{Markers and marker signs, seperated by commas}
}
}
\source{
\url{https://doi.org/10.1002/cyto.a.23725}
}
\usage{
Dusoswa_OMIP_54_markers
}
\description{
This data was originally published by Dusoswa et al. in 2019.
Marker definitions are extracted from the published Optimized Multicolor Immunofluorescence Panel (OMIP), which was designed for studies using mice brain, spleen, and bone marrow samples on the mass cytometer.
}
\keyword{datasets}
