% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kimmey_5hr_stim}
\alias{Kimmey_5hr_stim}
\title{Kimmey_5hr_stim}
\format{
A data frame with 95492 rows and 12 variables:

\describe{
\item{Cluster}{Cluster the cell was grouped into (1-9)}
\item{CD16}{}
\item{CD56}{}
\item{CD123}{}
\item{CD4}{}
\item{CD20}{}
\item{CD14}{}
\item{CD19}{}
\item{CD38}{}
\item{CD3}{}
\item{CD11c}{}
\item{HLA.DR}{}
}
}
\source{
\url{https://doi.org/10.1038/s41467-019-09128-7a}
}
\usage{
Kimmey_5hr_stim
}
\description{
This data was originally published by Kimmey et al. in 2019.
It is human PBMC mass cytometry data from a single donor.
The sample underwent a 5-hour phorbol 12-myristate 13-acetate (PMA) ionomycin stimulation prior to being inputted into the cytometer.
}
\keyword{datasets}
