% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lee_AML_cell_types_markers}
\alias{Lee_AML_cell_types_markers}
\title{Lee_AML_cell_types_markers}
\format{
A data frame with 14 rows and 2 columns:
\describe{
\item{Name}{Cell type name}
\item{Markers}{Markers and marker signs, seperated by commas}
}
}
\source{
\url{https://doi.org/10.1093/bioinformatics/btx054}
}
\usage{
Lee_AML_cell_types_markers
}
\description{
This cell type-marker table was originally published by Lee et al. in 2017 to test the Automated Cell-type Discovery and Classification (ACDC) algorithm.
That table was generated based on a gating hierarchy on Cytobank.
The original data was published by Levine et al. in 2015 and was collected from healthy human bone marrow.
}
\keyword{datasets}
