% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein_SPARQL.R
\name{PRO_SPARQL}
\alias{PRO_SPARQL}
\title{Function for SPARQL query and processing of the resulting data}
\usage{
PRO_SPARQL(unique_marker_names, NCBI_taxon_ID, onto_endpoint, Match_step)
}
\arguments{
\item{unique_marker_names}{A vector of marker names}

\item{NCBI_taxon_ID}{String specifying the NCBI taxon ID}

\item{onto_endpoint}{String specifying the Ontobee endpoint}

\item{Match_step}{String indicating step in the pipeline}
}
\value{
Dataframe of query results
}
\description{
Function for SPARQL query and processing of the resulting data
}
\examples{
unique_syn_marker_names = c("CD4", "CD16")
NCBI_taxon_ID_short = "NCBITaxon_9606"
onto_endpoint <- "https://sparql.hegroup.org/sparql"
Match_step <- "Step 2"
PRO_SPARQL(unique_marker_names = unique_syn_marker_names, NCBI_taxon_ID = NCBI_taxon_ID_short,
onto_endpoint = onto_endpoint, Match_step = Match_step)
}
