% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_app}
\alias{server_app}
\title{Shiny app server}
\arguments{
\item{input}{Object that contains all the input data sent from the browser, read only}

\item{output}{Object named according to the output ID}

\item{session}{Code run when session starts}
}
\value{
server
}
\description{
Shiny app server
}
